+++ CHANGES.md
@@ -1,5 +1,5 @@
 **Added**  
-* 
+* find references for global vars and commands across all files
 
 **Changed**  
 * 
@@ -8,7 +8,5 @@
 * 
 
 **Notes**
-* we need to find references for global vars and commands across all files
-    * we already have a reference contributor for global vars
 * fix renaming variable
 * scope works, but it doesn't work for deep nested scopes